html/semantics/embedded-content-0/media-elements/networkState_property_exists.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.networkState - existence test</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-networkstate">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_idl_attribute(
   document.getElementById("a"),
   "networkState",
   "audioElement.networkState property exists");
}, "audio.networkState - existence test");
test(function() {
 assert_idl_attribute(
   document.getElementById("v"),
   "networkState",
   "videoElement.networkState property exists");
}, "video.networkState - existence test");
  </script>
 </body>
</html>
